Fitting a model-based meta-analysis model using Stan

`MBMA_stan` fits a model-based meta-analysis model using Stan.

MBMA_stan(data = NULL, Pred_doses = NULL, model = "AB_Emax",
  mu_prior = c(0, 10), Emax_prior = c(0, 10), ED50_prior = c(0, 10),
  tau_prior = 0.5, tau_prior_dist = "half-normal", chains = 4,
  iter = 2000, warmup = 1000, adapt_delta = 0.95)

Arguments

data
An object of `create_MBMA_dat`.

Pred_doses
A numerical vector specfying the doses which prediction will be made. Default is NULL.

model
A string specifying the model used. Available options are `Baseline_Emax` (Baseline random effects model (Boucher and Bennets, 2016)), `CB_Emax` (contrast-based random effects model (Mawdsley et al ,2016)), `AB_Emax` (arm-based random effects model, adapted from (Zhang et al ,2017)), `CBPlusBaseline_Emax` (contrast-based plus baseline random effects model, adapted from (Dias et al ,2013)). Default is `AB_Emax`.

mu_prior
A numerical vector specifying the parameter of the normal prior density for baseline risks, first value is parameter for mean, second is for variance. Default is c(0, 10).

Emax_prior
A numerical vector specifying the parameter of the normal prior density for Emax parameter, first value is parameter for mean, second is for standard devation. Default is c(0, 10).

ED50_prior
A numerical vector specifying the parameter of the normal prior density for ED50 parameter, first value is parameter for mean, second is for standard devation. Default is c(0, 10).

tau_prior
A numerical value specifying the standard dev. of the prior density for heterogenety stdev. Default is 0.5.

tau_prior_dist
A string specifying the prior density for the heterogeneity standard deviation, option is `half-normal` for half-normal prior, `uniform` for uniform prior, `half-cauchy` for half-cauchy prior.

chains
A positive integer specifying the number of Markov chains. The default is 4.

iter
A positive integer specifying the number of iterations for each chain (including warmup). The default is 2000.

warmup
A positive integer specifying the number of warmup (aka burnin) iterations per chain. The default is 1000.

adapt_delta
A numerical value specfying the target average proposal acceptance probability for adaptation. See Stan manual for details. Default is 0.95. In general you should not need to change adapt_delta unless you see a warning message about divergent transitions, in which case you can increase adapt_delta from the default to a value closer to 1 (e.g. from 0.95 to 0.99, or from 0.99 to 0.999, etc).

Value

an object of class `stanfit` returned by `rstan::sampling`

References

Boucher M, et al. The many flavors of model-based meta-analysis: Part I-Introduction and landmark data. CPT: Pharmacometrics & Systems Pharmacology. 2016;5:54–64.

Mawdsley D, et al. Model-based network meta-analysis: A framework for evidence synthesis of clinical trial data. CPT: Pharmacometrics & Systems Pharmacology. 2016;5:393-401.

Zhang J, et al. (2014). Network meta-analysis of randomized clinical trials: Reporting the proper summaries. Clinical Trials. 11(2), 246–262.

Dias S, et al. Absolute or relative effects? Arm-based synthesis of trial data. Research Synthesis Methods. 2016;7:23–28.

Examples

## Load the dataset
data('dat.Eletriptan', package = "MetaStan")
## Fitting a Binomial-Normal Hierarchial model using WIP priors
datMBMA = create_MBMA_dat(dat = dat.Eletriptan,
                         armVars = c(dose = "d", responders = "r",
                                     sampleSize = "n"),
                         nArmsVar = "nd")

MBMA.AB.Emax  <- MBMA_stan(data = datMBMA,
                          model = "AB_Emax",
                          Pred_doses = seq(0, 80, length.out = 11),
                          Emax_prior = c(0, 10),
                          tau_prior_dist = "half-normal",
                          tau_prior = 0.5)

#> 
#> SAMPLING FOR MODEL 'AB_Emax' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000414 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 4.14 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 4.52039 seconds (Warm-up)
#> Chain 1:                3.51057 seconds (Sampling)
#> Chain 1:                8.03096 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'AB_Emax' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 9.5e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.95 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 4.19691 seconds (Warm-up)
#> Chain 2:                3.61871 seconds (Sampling)
#> Chain 2:                7.81563 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'AB_Emax' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 6.9e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.69 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 4.2156 seconds (Warm-up)
#> Chain 3:                3.60494 seconds (Sampling)
#> Chain 3:                7.82054 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'AB_Emax' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000114 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.14 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 4.44503 seconds (Warm-up)
#> Chain 4:                3.55418 seconds (Sampling)
#> Chain 4:                7.99921 seconds (Total)
#> Chain 4: