Prepare model-based meta-analysis dataset for Stan.

create_MBMA_dat converts datasets in the one-study-per-row format to one-arm-per-row format,

create_MBMA_dat(dat = dat, armVars = c(dose = "t", responders = "r",
  sampleSize = "n"), nArmsVar = "nd")

Arguments

dat
Data in one-study-per-row format.

armVars
Vector of per-arm variables The name of each component will be the column name in the resulting dataset.

nArmsVar
Variable holding the number of arms for each study.

Value

A data frame with the generated coloumns.

Details

The resulting data.frame can be used as data argument in MBMA_stan.

See also

Examples

data('dat.Eletriptan')
## Create the dataset suitable for MBMA_stan
EletriptanDat <- create_MBMA_dat(dat = dat.Eletriptan,
armVars = c("dose" = "d", "r" = "r", "n" = "n"), nArmsVar = 'nd')
## Check that the data are correct
print(EletriptanDat)

#>    study dose   r   n na
#> 1      1    0   6  70  2
#> 2      1   40  24  69  2
#> 3      2    0  43 195  3
#> 4      2   40  93 197  3
#> 5      2   20  61 173  3
#> 6      3    0   8 135  3
#> 7      3   80 157 360  3
#> 8      3   40 115 359  3
#> 9      4    0  17  89  2
#> 10     4   40  64 184  2
#> 11     5    0   5  74  3
#> 12     5   80  64 152  3
#> 13     5   40  54 154  3
#> 14     6    0   8 168  3
#> 15     6   80  93 401  3
#> 16     6   40 140 431  3
#> 17     7    0  20 404  2
#> 18     7   40 280 779  2
#> 19     8    0  12 292  4
#> 20     8   40  80 296  4
#> 21     8   80  84 312  4
#> 22     8   20  41 290  4
#> 23     9    0  10  79  4
#> 24     9   40  17  76  4
#> 25     9   80  21  74  4
#> 26     9   20  19  80  4
#> 27    10    0   7 232  3
#> 28    10   80 152 446  3
#> 29    10   40 138 430  3
#> 30    11    0   8 126  2
#> 31    11   20  25 129  2
#> 32    12    0   5 102  3
#> 33    12   80  79 209  3
#> 34    12   40  58 206  3