Prepare model-based meta-analysis dataset for Stan.

create_MBMA_dat converts datasets in the one-study-per-row format to one-arm-per-row format,

create_MBMA_dat(
  dat = dat,
  armVars = c(dose = "d", r = "r", n = "n"),
  nArmsVar = "nd"
)

Arguments

dat
Data in one-study-per-row format.

armVars
Vector of per-arm variables The name of each component will be the column name in the resulting dataset.

nArmsVar
Variable holding the number of arms for each study.

Value

A data frame with the generated coloumns.

Details

The resulting data.frame can be used as data argument in MBMA_stan.

See also

Author

Burak Kuersad Guenhan, burak.gunhan@med.uni-goettingen.de and Gert van Valkenhoef

Examples

if (FALSE) {
data('dat.Eletriptan')
## Create the dataset suitable for MBMA_stan
EletriptanDat <- create_MBMA_dat(dat = dat.Eletriptan,
                                 armVars = c("dose" = "d", "r" = "r", "n" = "n"),
                                 nArmsVar = 'nd')
## Check that the data are correct
print(EletriptanDat)
}